+++ src/main/java/fr/maxlego08/essentials/module/ZModuleManager.java
@@ -13,6 +13,8 @@ import fr.maxlego08.essentials.module.modules.chat.ChatModule;
 import fr.maxlego08.essentials.module.modules.customcommands.CustomCommandModule;
 import dev.yanianz.essentials.customscreens.CustomScreenModule;
 import dev.yanianz.essentials.polls.PollsModule;
+import dev.yanianz.essentials.nicknames.NicknamesModule;
+import dev.yanianz.essentials.reports.ReportsModule;
 import dev.yanianz.essentials.notes.NotesModule;
 import dev.yanianz.essentials.reputation.ReputationModule;
 import fr.maxlego08.essentials.module.modules.effects.EffectsModule;
@@ -97,6 +99,8 @@ public class ZModuleManager implements ModuleManager {
         this.modules.put(PollsModule.class, new PollsModule(this.plugin));
         this.modules.put(ReputationModule.class, new ReputationModule(this.plugin));
         this.modules.put(NotesModule.class, new NotesModule(this.plugin));
+        this.modules.put(NicknamesModule.class, new NicknamesModule(this.plugin));
+        this.modules.put(ReportsModule.class, new ReportsModule(this.plugin));
         if (plugin.isPaperVersion()) {
             this.modules.put(DeathMessageModule.class, new DeathMessageModule(this.plugin));
         }